requirements.txt
numpy
scipy
librosa
ffmpeg-python
soundfile
streamlit_vertical_slider
streamlit_toggle
soundfile
pydub
python-wordpress-xmlrpc
woocommerce
altair
pandas
pydeck
ffmpeg
streamlit-extras
streamlit_toggle


torch
torchaudio 
transformers 
gradio
PyJWT==2.3.0
extra-streamlit-components
